Changed methodology of saving the active station location of user
Previously the active station location was determined through the station_profiles table and the active station was marked. This is ok if a user will just see its own locations but for station location sharing between users this won't work anymore. So the column active_station_location was added to users table to store this information and the most obvious calls were adapted to it. The old columns are still maintained for the case that they are used somewhere else.