Made QSO logging possible into locations not owned but shared to a user
Added function check_station_is_writable() to model Stations.php to check if user has write permissions in station location and used this one insteead of check_station_is_accessible() for addition of a log in the database. Used a new function instead of the old one to make it easier to implement finer grained role model for shares later on.